원거리 공격 재수정

+++ Assets/Script/PlaceMonster.cs
@@ -442,8 +442,7 @@ public class PlaceMonster : MonoBehaviour {
 
 
             int range = to - from;
-            if (to != -1) {
-                if (targetList.Count - 1 < from) range = 0;
+            if (to != -1 && from < targetList.Count - 1) {
                 targetList.RemoveRange(from, range);
             }
         }
